requirements.txt
Flask==2.1.3
numpy<1.23.1
Pillow==9.2.0
tensorflow-cpu
gunicorn